Runtime Symbols V2 + Cleanup

+++ BlamboozleLib/halo5/halo5_forge.h
@@ -1,9 +1,9 @@
 #pragma once
 
-static constexpr uintptr_t h5_base_address = 0x0000000140000000;
-//static constexpr uintptr_t h5_tag_layout_table_address = 0x000000014477A2D0; // _build_halo5_forge_1_114_4592_2
+static constexpr uint64_t h5_base_address = 0x0000000140000000;
+//static constexpr uint64_t h5_tag_layout_table_address = 0x000000014477A2D0; // _build_halo5_forge_1_114_4592_2
 // static constexpr uint32_t h5_num_tag_layouts = 284; // _build_halo5_forge_1_114_4592_2
-static constexpr uintptr_t h5_tag_layout_table_address = 0x0000000144756580; // _build_halo5_forge_1_194_6192_2
+static constexpr uint64_t h5_tag_layout_table_address = 0x0000000144756580; // _build_halo5_forge_1_194_6192_2
 static constexpr uint32_t h5_num_tag_layouts = 285; // _build_halo5_forge_1_194_6192_2
 
 // #TODO: More sophisticated addressing using sections
